pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "a-brick-translator"
version = "1.0.0"
description = "A smart translation tool supporting multiple AI models"
readme = "README.md"
authors = [
    {name = "koucxz", email = "koucxz@hotmail.com"}
]
license = {text = "MIT"}
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "anthropic>=0.18.0",
    "dashscope>=1.14.0",
    "openai>=1.0.0",
    "google-genai>=1.0.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "autopep8>=2.0.0",
    "pyyaml>=6.0.0"
]
requires-python = ">=3.8"

[project.urls]
Homepage = "https://github.com/koucxz/a-brick-translator"
Repository = "https://github.com/koucxz/a-brick-translator"

[project.scripts]
brick-translator = "brick_translator.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["brick_translator*"]